fix printf formats; use uint32_t or uint64_t instead of unsigned

+++ include/RevMemCtrl.h
@@ -83,30 +83,30 @@ void RevHandleFlagResp( void* target, size_t size, RevFlag flags );
 class RevMemOp {
 public:
   /// RevMemOp constructor
-  RevMemOp( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, MemOp Op, RevFlag flags );
+  RevMemOp( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, MemOp Op, RevFlag flags );
 
   /// RevMemOp constructor
-  RevMemOp( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, MemOp Op, RevFlag flags );
+  RevMemOp( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, MemOp Op, RevFlag flags );
 
   /// RevMemOp overloaded constructor
-  RevMemOp( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, MemOp Op, RevFlag flags );
+  RevMemOp( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, MemOp Op, RevFlag flags );
 
   /// RevMemOp overloaded constructor
   RevMemOp(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, void* target, MemOp Op, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, void* target, MemOp Op, RevFlag flags
   );
 
   /// RevMemOp overloaded constructor
-  RevMemOp( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, std::vector<uint8_t> buffer, MemOp Op, RevFlag flags );
+  RevMemOp( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, std::vector<uint8_t> buffer, MemOp Op, RevFlag flags );
 
   /// RevMemOp overloaded constructor
   RevMemOp(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, unsigned CustomOpc, MemOp Op, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, uint32_t CustomOpc, MemOp Op, RevFlag flags
   );
 
   /// RevMemOp overloaded constructor
   RevMemOp(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, unsigned CustomOpc, MemOp Op, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, uint32_t CustomOpc, MemOp Op, RevFlag flags
   );
 
   /// RevMemOp default destructor
@@ -120,7 +120,7 @@ public:
   MemOp getOp() const { return Op; }
 
   /// RevMemOp: retrieve the custom opcode
-  unsigned getCustomOpc() const { return CustomOpc; }
+  uint32_t getCustomOpc() const { return CustomOpc; }
 
   /// RevMemOp: retrieve the target address
   uint64_t getAddr() const { return Addr; }
@@ -147,13 +147,13 @@ public:
   RevFlag getNonCacheFlags() const { return RevFlag{ safe_static_cast<uint32_t>( flags ) & 0xFFFD }; }
 
   /// RevMemOp: sets the number of split cache line requests
-  void setSplitRqst( unsigned S ) { SplitRqst = S; }
+  void setSplitRqst( uint32_t S ) { SplitRqst = S; }
 
   /// RevMemOp: set the invalidate flag
   void setInv( bool I ) { Inv = I; }
 
   /// RevMemOp: set the hart
-  void setHart( unsigned H ) { Hart = H; }
+  void setHart( uint32_t H ) { Hart = H; }
 
   /// RevMemOp: set the originating memory request
   void setMemReq( const MemReq& req ) { procReq = req; }
@@ -165,13 +165,13 @@ public:
   bool getInv() const { return Inv; }
 
   /// RevMemOp: retrieve the number of split cache line requests
-  unsigned getSplitRqst() const { return SplitRqst; }
+  uint32_t getSplitRqst() const { return SplitRqst; }
 
   /// RevMemOp: retrieve the target address
   void* getTarget() const { return target; }
 
   /// RevMemOp: retrieve the hart
-  unsigned getHart() const { return Hart; }
+  uint32_t getHart() const { return Hart; }
 
   /// RevMemOp: Get the originating proc memory request
   const MemReq& getMemReq() const { return procReq; }
@@ -182,14 +182,14 @@ public:
   }
 
 private:
-  unsigned             Hart{};       ///< RevMemOp: RISC-V Hart
+  uint32_t             Hart{};       ///< RevMemOp: RISC-V Hart
   uint64_t             Addr{};       ///< RevMemOp: address
   uint64_t             PAddr{};      ///< RevMemOp: physical address (for RevMem I/O)
   uint32_t             Size{};       ///< RevMemOp: size of the memory operation in bytes
   bool                 Inv{};        ///< RevMemOp: flush operation invalidate flag
   MemOp                Op{};         ///< RevMemOp: target memory operation
-  unsigned             CustomOpc{};  ///< RevMemOp: custom memory opcode
-  unsigned             SplitRqst{};  ///< RevMemOp: number of split cache line requests
+  uint32_t             CustomOpc{};  ///< RevMemOp: custom memory opcode
+  uint32_t             SplitRqst{};  ///< RevMemOp: number of split cache line requests
   std::vector<uint8_t> membuf{};     ///< RevMemOp: buffer
   std::vector<uint8_t> tempT{};      ///< RevMemOp: temporary target buffer for R-M-W ops
   RevFlag              flags{};      ///< RevMemOp: request flags
@@ -217,7 +217,7 @@ public:
   RevMemCtrl& operator=( const RevMemCtrl& )                                                                            = delete;
 
   /// RevMemCtrl: initialization function
-  void init( unsigned int phase ) override                                                                              = 0;
+  void init( uint32_t phase ) override                                                                                  = 0;
 
   /// RevMemCtrl: setup function
   void setup() override                                                                                                 = 0;
@@ -229,20 +229,20 @@ public:
   virtual bool outstandingRqsts()                                                                                       = 0;
 
   /// RevMemCtrl: send flush request
-  virtual bool sendFLUSHRequest( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, bool Inv, RevFlag flags ) = 0;
+  virtual bool sendFLUSHRequest( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, bool Inv, RevFlag flags ) = 0;
 
   /// RevMemCtrl: send a read request
   virtual bool sendREADRequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, const MemReq& req, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, const MemReq& req, RevFlag flags
   ) = 0;
 
   /// RevMemCtrl: send a write request
   virtual bool
-    sendWRITERequest( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) = 0;
+    sendWRITERequest( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) = 0;
 
   /// RevMemCtrl: send an AMO request
   virtual bool sendAMORequest(
-    unsigned       Hart,
+    uint32_t       Hart,
     uint64_t       Addr,
     uint64_t       PAddr,
     uint32_t       Size,
@@ -254,32 +254,32 @@ public:
 
   /// RevMemCtrl: send a readlock request
   virtual bool sendREADLOCKRequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, const MemReq& req, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, const MemReq& req, RevFlag flags
   ) = 0;
 
   /// RevMemCtrl: send a writelock request
   virtual bool
-    sendWRITELOCKRequest( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) = 0;
+    sendWRITELOCKRequest( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) = 0;
 
   /// RevMemCtrl: send a loadlink request
-  virtual bool sendLOADLINKRequest( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, RevFlag flags )              = 0;
+  virtual bool sendLOADLINKRequest( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, RevFlag flags )              = 0;
 
   /// RevMemCtrl: send a storecond request
   virtual bool
-    sendSTORECONDRequest( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) = 0;
+    sendSTORECONDRequest( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) = 0;
 
   /// RevMemCtrl: send an void custom read memory request
   virtual bool sendCUSTOMREADRequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, unsigned Opc, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, uint32_t Opc, RevFlag flags
   ) = 0;
 
   /// RevMemCtrl: send a custom write request
   virtual bool sendCUSTOMWRITERequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, unsigned Opc, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, uint32_t Opc, RevFlag flags
   )                                                            = 0;
 
   /// RevMemCtrl: send a FENCE request
-  virtual bool sendFENCE( unsigned Hart )                      = 0;
+  virtual bool sendFENCE( uint32_t Hart )                      = 0;
 
   /// RevMemCtrl: handle a read response
   virtual void handleReadResp( StandardMem::ReadResp* ev )     = 0;
@@ -303,7 +303,7 @@ public:
   virtual void handleAMO( RevMemOp* op )                       = 0;
 
   /// RevMemCtrl: returns the cache line size
-  virtual unsigned getLineSize()                               = 0;
+  virtual uint32_t getLineSize()                               = 0;
 
   /// Assign processor tracer
   virtual void setTracer( RevTracer* tracer )                  = 0;
@@ -444,7 +444,7 @@ public:
   RevBasicMemCtrl& operator=( const RevBasicMemCtrl& ) = delete;
 
   /// RevBasicMemCtrl: initialization function
-  void init( unsigned int phase ) final;
+  void init( uint32_t phase ) final;
 
   /// RevBasicMemCtrl: setup function
   void setup() final;
@@ -459,27 +459,27 @@ public:
   bool outstandingRqsts() final;
 
   /// RevBasicMemCtrl: returns the cache line size
-  unsigned getLineSize() final { return lineSize; }
+  uint32_t getLineSize() final { return lineSize; }
 
   /// RevBasicMemCtrl: memory event processing handler
   void processMemEvent( StandardMem::Request* ev );
 
   /// RevBasicMemCtrl: send a flush request
-  bool sendFLUSHRequest( unsigned Hart, uint64_t Addr, uint64_t PAdr, uint32_t Size, bool Inv, RevFlag flags ) final;
+  bool sendFLUSHRequest( uint32_t Hart, uint64_t Addr, uint64_t PAdr, uint32_t Size, bool Inv, RevFlag flags ) final;
 
   /// RevBasicMemCtrl: send a read request
   bool sendREADRequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, const MemReq& req, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, const MemReq& req, RevFlag flags
   ) final;
 
   /// RevBasicMemCtrl: send a write request
   bool sendWRITERequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags = RevFlag::F_NONE
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags = RevFlag::F_NONE
   ) final;
 
   /// RevBasicMemCtrl: send an AMO request
   bool sendAMORequest(
-    unsigned       Hart,
+    uint32_t       Hart,
     uint64_t       Addr,
     uint64_t       PAddr,
     uint32_t       Size,
@@ -491,32 +491,32 @@ public:
 
   // RevBasicMemCtrl: send a readlock request
   bool sendREADLOCKRequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, const MemReq& req, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, const MemReq& req, RevFlag flags
   ) final;
 
   // RevBasicMemCtrl: send a writelock request
   bool
-    sendWRITELOCKRequest( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) final;
+    sendWRITELOCKRequest( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) final;
 
   // RevBasicMemCtrl: send a loadlink request
-  bool sendLOADLINKRequest( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, RevFlag flags ) final;
+  bool sendLOADLINKRequest( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, RevFlag flags ) final;
 
   // RevBasicMemCtrl: send a storecond request
   bool
-    sendSTORECONDRequest( unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) final;
+    sendSTORECONDRequest( uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, RevFlag flags ) final;
 
   // RevBasicMemCtrl: send an void custom read memory request
   bool sendCUSTOMREADRequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, unsigned Opc, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, void* target, uint32_t Opc, RevFlag flags
   ) final;
 
   // RevBasicMemCtrl: send a custom write request
   bool sendCUSTOMWRITERequest(
-    unsigned Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, unsigned Opc, RevFlag flags
+    uint32_t Hart, uint64_t Addr, uint64_t PAddr, uint32_t Size, unsigned char* buffer, uint32_t Opc, RevFlag flags
   ) final;
 
   // RevBasicMemCtrl: send a FENCE request
-  bool sendFENCE( unsigned Hart ) final;
+  bool sendFENCE( uint32_t Hart ) final;
 
   /// RevBasicMemCtrl: handle a read response
   void handleReadResp( StandardMem::ReadResp* ev ) final;
@@ -583,26 +583,26 @@ protected:
 private:
   /// RevBasicMemCtrl: process the next memory request
   bool processNextRqst(
-    unsigned& t_max_loads,
-    unsigned& t_max_stores,
-    unsigned& t_max_flush,
-    unsigned& t_max_llsc,
-    unsigned& t_max_readlock,
-    unsigned& t_max_writeunlock,
-    unsigned& t_max_custom,
-    unsigned& t_max_ops
+    uint32_t& t_max_loads,
+    uint32_t& t_max_stores,
+    uint32_t& t_max_flush,
+    uint32_t& t_max_llsc,
+    uint32_t& t_max_readlock,
+    uint32_t& t_max_writeunlock,
+    uint32_t& t_max_custom,
+    uint32_t& t_max_ops
   );
 
   /// RevBasicMemCtrl: determine if we can instantiate the target memory operation
   bool isMemOpAvail(
     RevMemOp* Op,
-    unsigned& t_max_loads,
-    unsigned& t_max_stores,
-    unsigned& t_max_flush,
-    unsigned& t_max_llsc,
-    unsigned& t_max_readlock,
-    unsigned& t_max_writeunlock,
-    unsigned& t_max_custom
+    uint32_t& t_max_loads,
+    uint32_t& t_max_stores,
+    uint32_t& t_max_flush,
+    uint32_t& t_max_llsc,
+    uint32_t& t_max_readlock,
+    uint32_t& t_max_writeunlock,
+    uint32_t& t_max_custom
   );
 
   /// RevBasicMemCtrl: build a standard memory request
@@ -615,13 +615,13 @@ private:
   bool buildCacheMemRqst( RevMemOp* op, bool& Success );
 
   /// RevBasicMemCtrl: determine if there are any pending AMOs that would prevent a request from dispatching
-  bool isPendingAMO( unsigned Slot );
+  bool isPendingAMO( uint32_t Slot );
 
   /// RevBasicMemCtrl: determine if we need to utilize AQ ordering semantics
-  bool isAQ( unsigned Slot, unsigned Hart );
+  bool isAQ( uint32_t Slot, uint32_t Hart );
 
   /// RevBasicMemCtrl: determine if we need to utilize RL ordering semantics
-  bool isRL( unsigned Slot, unsigned Hart );
+  bool isRL( uint32_t Slot, uint32_t Hart );
 
   /// RevBasicMemCtrl: register statistics
   void registerStats();
@@ -633,30 +633,30 @@ private:
   uint64_t getTotalRqsts();
 
   /// RevBasicMemCtrl: Determine the number of cache lines are required
-  unsigned getNumCacheLines( uint64_t Addr, uint32_t Size );
+  uint32_t getNumCacheLines( uint64_t Addr, uint32_t Size );
 
   /// RevBasicMemCtrl: Retrieve the base cache line request size
-  unsigned getBaseCacheLineSize( uint64_t Addr, uint32_t Size );
+  uint32_t getBaseCacheLineSize( uint64_t Addr, uint32_t Size );
 
   /// RevBasicMemCtrl: retrieve the number of outstanding requests on the wire
-  unsigned getNumSplitRqsts( RevMemOp* op );
+  uint32_t getNumSplitRqsts( RevMemOp* op );
 
   /// RevBasicMemCtrl: perform the MODIFY portion of the AMO (READ+MODIFY+WRITE)
-  void performAMO( std::tuple<unsigned, unsigned char*, void*, RevFlag, RevMemOp*, bool> Entry );
+  void performAMO( std::tuple<uint32_t, unsigned char*, void*, RevFlag, RevMemOp*, bool> Entry );
 
   // -- private data members
   StandardMem*       memIface{};         ///< StandardMem memory interface
   RevStdMemHandlers* stdMemHandlers{};   ///< StandardMem interface response handlers
   bool               hasCache{};         ///< detects whether cache layers are present
-  unsigned           lineSize{};         ///< cache line size
-  unsigned           max_loads{};        ///< maximum number of outstanding loads
-  unsigned           max_stores{};       ///< maximum number of outstanding stores
-  unsigned           max_flush{};        ///< maximum number of oustanding flush events
-  unsigned           max_llsc{};         ///< maximum number of outstanding llsc events
-  unsigned           max_readlock{};     ///< maximum number of oustanding readlock events
-  unsigned           max_writeunlock{};  ///< maximum number of oustanding writelock events
-  unsigned           max_custom{};       ///< maximum number of oustanding custom events
-  unsigned           max_ops{};          ///< maximum number of ops to issue per cycle
+  uint64_t           lineSize{};         ///< cache line size
+  uint64_t           max_loads{};        ///< maximum number of outstanding loads
+  uint64_t           max_stores{};       ///< maximum number of outstanding stores
+  uint64_t           max_flush{};        ///< maximum number of oustanding flush events
+  uint64_t           max_llsc{};         ///< maximum number of outstanding llsc events
+  uint64_t           max_readlock{};     ///< maximum number of oustanding readlock events
+  uint64_t           max_writeunlock{};  ///< maximum number of oustanding writelock events
+  uint64_t           max_custom{};       ///< maximum number of oustanding custom events
+  uint64_t           max_ops{};          ///< maximum number of ops to issue per cycle
 
   uint64_t num_read{};         ///< number of outstanding read requests
   uint64_t num_write{};        ///< number of outstanding write requests
@@ -679,7 +679,7 @@ private:
 #define AMOTABLE_IN     5
 
   /// RevBasicMemCtrl: map of amo operations to memory addresses
-  std::multimap<uint64_t, std::tuple<unsigned, unsigned char*, void*, RevFlag, RevMemOp*, bool>> AMOTable{};
+  std::multimap<uint64_t, std::tuple<uint32_t, unsigned char*, void*, RevFlag, RevMemOp*, bool>> AMOTable{};
 
   std::vector<Statistic<uint64_t>*> stats{};  ///< statistics vector
